refactor: move some packages into internal
