+++ app/server.go
@@ -13,9 +13,10 @@ import (
 	guardianv1beta1 "github.com/odpf/guardian/api/proto/odpf/guardian/v1beta1"
 	"github.com/odpf/guardian/appeal"
 	"github.com/odpf/guardian/approval"
-	"github.com/odpf/guardian/crypto"
 	"github.com/odpf/guardian/domain"
 	"github.com/odpf/guardian/iam"
+	"github.com/odpf/guardian/internal/crypto"
+	"github.com/odpf/guardian/internal/scheduler"
 	"github.com/odpf/guardian/model"
 	"github.com/odpf/guardian/notifier"
 	"github.com/odpf/guardian/policy"
@@ -26,7 +27,6 @@ import (
 	"github.com/odpf/guardian/provider/metabase"
 	"github.com/odpf/guardian/provider/tableau"
 	"github.com/odpf/guardian/resource"
-	"github.com/odpf/guardian/scheduler"
 	"github.com/odpf/guardian/store"
 	"github.com/odpf/salt/log"
 	"golang.org/x/net/http2"
