+++ approval/service.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 
 	"github.com/odpf/guardian/domain"
-	"github.com/odpf/guardian/evaluator"
+	"github.com/odpf/guardian/internal/evaluator"
 )
 
 type service struct {
