+++ policy/service.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/go-playground/validator/v10"
 	"github.com/odpf/guardian/domain"
-	"github.com/odpf/guardian/evaluator"
+	"github.com/odpf/guardian/internal/evaluator"
 )
 
 // Service handling the business logics
